base code WORKING
worked together with francis for this. we locked ourselves in the hfsc to finish the base code and functions for our web app, mainly using my computer. we only committed when everything finally worked correctly after a lot of trial and error because our website kept breaking and we didn't want to upload broken code.